feat(data-sync): Phase 7 - extract SyncEvent types + finalize (EW-628)
Splits the SyncEvent / SyncEventSource discriminated-union types out of
SyncEventRow.tsx into a sibling SyncEvent.types.ts so they can be
consumed by upcoming surface code (Phase 7 follow-up FeedRow adapter,
API-client glue translating raw activity-log rows into the
SyncEvent union, server-side fixtures) without pulling the component
implementation. SSR/test consumers stay tree-shake friendly.

Re-exports from SyncEventRow.tsx preserve the existing import path so
the unit tests (and any future caller that already imported from the
component file) keep working without changes.

Promotes b723e5c7 wip to feat — Phase 7 is complete for the
component-and-types scope. FeedFilterChips integration (extending
FEED_CATEGORIES with 'data-sync' + the API-client adapter) lands on
the Phase 7 follow-up so the type cascade through the existing
activity feed surface stays isolated and reviewable.

Refs EW-628.
